cleaned up new warning for unused variables Add a new method SWBuf::indexOf
git-svn-id: https://crosswire.org/svn/sword/trunk@2659 bcd7d363-81e1-0310-97ec-a550e20fc99c

+++ tests/ciphertest.cpp
@@ -30,7 +30,7 @@ int main(int argc, char **argv) {
 	}
 
 	
-	int encipher = atoi(argv[2]);
+	long encipher = atoi(argv[2]);
 
 	SWFilter *filter = new CipherFilter(argv[1]);
